Revert "Add WIP support for customized mod generation"
This reverts commit 421bab316ceaacbf87961a33847f32c86a35299f.
