+++ templates/architectury/common/src/main/resources/architectury.common.json
@@ -1,3 +1,3 @@
 {
-  "accessWidener": "@MODID@.accesswidener"
+  "accessWidener": "examplemod.accesswidener"
 }
