--- templates/architectury/common/src/main/resources/assets/@MODID@/lang/en_us.json
@@ -1,3 +1,0 @@
-{
-  "item.@MODID@.example_item": "Example Item"
-}
