+++ templates/architectury/common/src/main/resources/assets/examplemod/lang/en_us.json
@@ -1,0 +1,3 @@
+{
+  "item.examplemod.example_item": "Example Item"
+}
