+++ templates/architectury/common/src/main/resources/examplemod.accesswidener
