--- templates/architectury/forge/src/main/java/@PACKAGE@/forge/ExampleExpectPlatformImpl.java
@@ -1,15 +1,0 @@
-package @PACKAGE@.forge;
-
-import @PACKAGE@.ExampleExpectPlatform;
-import net.minecraftforge.fml.loading.FMLPaths;
-
-import java.nio.file.Path;
-
-public class ExampleExpectPlatformImpl {
-    /**
-     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
-     */
-    public static Path getConfigDirectory() {
-        return FMLPaths.CONFIGDIR.get();
-    }
-}
