--- templates/architectury/forge/src/main/java/@PACKAGE@/forge/ExampleModForge.java
@@ -1,15 +1,0 @@
-package @PACKAGE@.forge;
-
-import @ARCHITECTURY_PACKAGE@.platform.forge.EventBuses;
-import @PACKAGE@.ExampleMod;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-
-@Mod(ExampleMod.MOD_ID)
-public class ExampleModForge {
-    public ExampleModForge() {
-        // Submit our event bus to let architectury register our content on the right time
-        EventBuses.registerModEventBus(ExampleMod.MOD_ID, FMLJavaModLoadingContext.get().getModEventBus());
-        ExampleMod.init();
-    }
-}
