+++ templates/architectury/gradle.properties
@@ -3,9 +3,9 @@ org.gradle.jvmargs=-Xmx2048M
 minecraft_version=@MINECRAFT@
 enabled_platforms=@COMMON_MOD_LOADERS@
 
-archives_base_name=@MODID@
+archives_base_name=examplemod
 mod_version=1.0.0
-maven_group=@PACKAGE@
+maven_group=net.examplemod
 
 architectury_version=@ARCHITECTURY_API@
