--- templates/architectury_fabric/fabric/src/main/java/@PACKAGE/fabric/ExampleModFabric.java
@@ -1,11 +1,0 @@
-package @PACKAGE@.fabric;
-
-import @PACKAGE@.ExampleMod;
-import net.fabricmc.api.ModInitializer;
-
-public class ExampleModFabric implements ModInitializer {
-    @Override
-    public void onInitialize() {
-        ExampleMod.init();
-    }
-}
