--- templates/architectury_fabric_like/fabric-like/src/main/java/@PACKAGE@/fabriclike/ExampleModFabricLike.java
@@ -1,9 +1,0 @@
-package @PACKAGE@.fabriclike;
-
-import @PACKAGE@.ExampleMod;
-
-public class ExampleModFabricLike {
-    public static void init() {
-        ExampleMod.init();
-    }
-}
