--- templates/architectury_fabric_like/fabric/src/main/java/@PACKAGE@/fabric/ExampleModFabric.java
@@ -1,11 +1,0 @@
-package @PACKAGE@.fabric;
-
-import @PACKAGE@.fabriclike.ExampleModFabricLike;
-import net.fabricmc.api.ModInitializer;
-
-public class ExampleModFabric implements ModInitializer {
-    @Override
-    public void onInitialize() {
-        ExampleModFabricLike.init();
-    }
-}
