+++ templates/architectury_fabric_like/fabric/src/main/java/net/examplemod/fabric/ExampleModFabric.java
@@ -1,0 +1,11 @@
+package net.examplemod.fabric;
+
+import net.examplemod.fabriclike.ExampleModFabricLike;
+import net.fabricmc.api.ModInitializer;
+
+public class ExampleModFabric implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        ExampleModFabricLike.init();
+    }
+}
