--- templates/architectury_fabric_like/quilt/src/main/java/@PACKAGE@/quilt/ExampleModQuilt.java
@@ -1,12 +1,0 @@
-package @PACKAGE@.quilt;
-
-import @PACKAGE@.fabriclike.ExampleModFabricLike;
-import org.quiltmc.loader.api.ModContainer;
-import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
-
-public class ExampleModQuilt implements ModInitializer {
-    @Override
-    public void onInitialize(ModContainer mod) {
-        ExampleModFabricLike.init();
-    }
-}
