+++ templates/architectury_fabric_like/quilt/src/main/java/net/examplemod/quilt/ExampleModQuilt.java
@@ -1,0 +1,12 @@
+package net.examplemod.quilt;
+
+import net.examplemod.fabriclike.ExampleModFabricLike;
+import org.quiltmc.loader.api.ModContainer;
+import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
+
+public class ExampleModQuilt implements ModInitializer {
+    @Override
+    public void onInitialize(ModContainer mod) {
+        ExampleModFabricLike.init();
+    }
+}
