+++ templates/architectury_quilt/quilt/src/main/java/net/examplemod/fabric/ExampleExpectPlatformImpl.java
@@ -1,0 +1,15 @@
+package net.examplemod.fabric;
+
+import net.examplemod.ExampleExpectPlatform;
+import org.quiltmc.loader.api.QuiltLoader;
+
+import java.nio.file.Path;
+
+public class ExampleExpectPlatformImpl {
+    /**
+     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
+     */
+    public static Path getConfigDirectory() {
+        return QuiltLoader.getConfigDir();
+    }
+}
