+++ templates/forge/gradle.properties
@@ -17,7 +17,7 @@ loom.platform=forge
 
 # Mod Properties
     mod_version=1.0.0
-    maven_group=@PACKAGE@
-    archives_base_name=@MODID@
-    mod_id=@MODID@
+    maven_group=com.example.examplemod
+    archives_base_name=archloom-example-mod
+    mod_id=archloom-example-mod
     mod_author=TheExampleDev
