--- templates/forge/src/main/java/@PACKAGE@/mixin/ExampleMixin.java
@@ -1,16 +1,0 @@
-package @PACKAGE@.mixin;
-
-import @PACKAGE@.ExampleMod;
-import net.minecraft.client.gui.screens.TitleScreen;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(TitleScreen.class)
-public class ExampleMixin {
-    @Inject(at = @At("HEAD"), method = "init()V")
-    private void init(CallbackInfo info) {
-        ExampleMod.LOGGER.info("This line is printed by an example mod mixin!");
-    }
-}
