--- templates/forge/src/main/resources/@MODID@.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "package": "@PACKAGE@.mixin",
-  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
-  "mixins": [
-  ],
-  "client": [
-    "ExampleMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
