+++ templates/forge/src/main/resources/examplemod.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": true,
+  "package": "com.example.examplemod.mixin",
+  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
+  "mixins": [
+  ],
+  "client": [
+    "ExampleMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
