+++ templates/mixin/common/src/main/java/net/examplemod/mixin/MixinTitleScreen.java
@@ -1,0 +1,15 @@
+package net.examplemod.mixin;
+
+import net.minecraft.client.gui.screens.TitleScreen;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(TitleScreen.class)
+public class MixinTitleScreen {
+    @Inject(at = @At("HEAD"), method = "init()V")
+    private void init(CallbackInfo info) {
+        System.out.println("Hello from example architectury common mixin!");
+    }
+}
