--- templates/mixin/common/src/main/resources/@MODID@-common.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "package": "@PACKAGE@.mixin",
-  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
-  "client": [
-    "MixinTitleScreen"
-  ],
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
