+++ templates/mixin/common/src/main/resources/examplemod-common.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": true,
+  "package": "net.examplemod.mixin",
+  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
+  "client": [
+    "MixinTitleScreen"
+  ],
+  "mixins": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
