--- templates/mixin/forge/src/main/resources/@MODID@.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "@PACKAGE@.mixin.forge",
-  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
-  "client": [
-  ],
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
