+++ templates/mixin/forge/src/main/resources/examplemod.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": true,
+  "package": "net.examplemod.mixin.forge",
+  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
+  "client": [
+  ],
+  "mixins": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
