--- templates/mixin_fabric/fabric/src/main/resources/@MODID@.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "@PACKAGE@.mixin.fabric",
-  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
-  "client": [
-  ],
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
