+++ templates/mixin_quilt/quilt/src/main/resources/examplemod.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": true,
+  "package": "net.examplemod.mixin.fabric",
+  "compatibilityLevel": "@MIXIN_COMPAT_LEVEL@",
+  "client": [
+  ],
+  "mixins": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
